requirements.txt
APScheduler==3.10.1
prometheus-client==0.16.0
azure-mgmt-resource==21.1.0b1
azure-identity==1.12.0
azure-mgmt-compute==29.1.0